// src/app/components/card/card-create.dto.ts
export interface CreateCardDto {
  gameMode: string;
  championName: string;
  riotIdGameName?: string;
  riotIdTagline?: string;
  positionPlayer?: string;
  role?: string;
  realRole?: string;
  kills?: number;
  deaths?: number;
  assists?: number;
  kda?: string;
  killParticipation?: string;
  totalDamageDealtToChampions?: number;
  totalMinionsKilled?: number;
  totalNeutralMinionsKilled?: number;
  totalMinionsKilledJg?: number;
  teamId?: number;
  teamDragonsKilled?: number;
  teamBaronsKilled?: number;
  matchDragons?: number;
  matchBarons?: number;
  jungleKing?: boolean;
  gameLength?: number;
  damagePerMinute?: string;
  minionsPerMinute?: string;
  minionsPerMinuteJg?: string;
  goldPerMinute?: string;
  timeCCingOthers?: number;
  visionScore?: number;
  firstBloodKill?: boolean;
  firstBloodAssist?: boolean;
  firstTowerKill?: boolean;
  firstTowerAssist?: boolean;
  totalDamageShieldedOnTeammates?: number;
  totalHealsOnTeammates?: number;
  totalDamageTaken?: number;
  baronKills?: number;
  dragonKills?: number;
  quadraKills?: number;
  pentaKills?: number;
  splashArt?: string;
  iconChampion?: string;
  corDaBorda?: string;
  corDoVerso?: string;
  perks?: Record<string, string>;
  summonerSpells?: Record<string, string>;
  items?: Record<string, string | null>;
  augments?: Record<string, string | null>;
  achievements?: string[];
  gameDate?: string;
}